LayoutTests/imported/w3c/web-platform-tests/pointerevents/coalesced_events_attributes.https_mouse-expected.txt


FAIL Coalesced list in boundary events assert_equals: pointerover.getCoalescedEvents is a function expected "function" but got "undefined"
FAIL Coalesced list in pointer-capture events assert_equals: gotpointercapture.getCoalescedEvents is a function expected "function" but got "undefined"
FAIL Coalesced list in pointerdown/move/up events assert_equals: pointerdown.getCoalescedEvents is a function expected "function" but got "undefined"
PASS Coalesced list in pointercancel event
